int => short

+++ contact-management-system/contact-management-system/聯絡人管理系統.cpp
@@ -8,7 +8,7 @@ using namespace std;
 struct Person
 {
     string name; // 姓名
-    int sex = 0; // 性別 女 0 男 1　其他 2
+    short sex = 0; // 性別 女 0 男 1　其他 2
     short age = 18; // 年齡
     string phone; // 電話號碼
 };
